run_test.py for run test
