+++ tests/run_tests.py
@@ -1,3 +1,46 @@
+"""
+CalTrack Test Suite Runner
+
+This script allows you to run different types of tests for the CalTrack application.
+
+Usage:
+------
+1. Run all tests:
+   python -m tests.run_tests
+
+2. Run specific test categories:
+   - Unit tests (including calorie tracking tests):
+     python -m tests.run_tests unit
+   
+   - All Selenium tests:
+     python -m tests.run_tests selenium
+
+3. Run individual Selenium tests:
+   - Login page test:
+     python -m tests.run_tests login_page
+   
+   - Login success test:
+     python -m tests.run_tests login_success
+   
+   - Login failure test:
+     python -m tests.run_tests login_failure
+   
+   - Registration test:
+     python -m tests.run_tests registration
+   
+   - Track exercise test:
+     python -m tests.run_tests track_exercise
+   
+   - View profile test:
+     python -m tests.run_tests view_profile
+
+Notes:
+------
+- Unit tests are fast and don't require a browser
+- Selenium tests run in Chrome and test the full UI functionality
+- Selenium tests use random ports to avoid conflicts when running multiple tests
+"""
+
 import unittest
 import sys
 import os
@@ -6,8 +49,7 @@ import os
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 # Import test modules
-from tests.unitTest import CalTrackUnitTests
-from tests.test_calorie_tracking import CalorieTrackingTests
+from tests.unitTest import CalTrackUnitTests, CalorieTrackingTests
 from tests.seleniumTest import (
     TestLoginPage, 
     TestLoginSuccess, 
@@ -18,14 +60,13 @@ from tests.seleniumTest import (
 )
 
 def run_unit_tests():
-    """Run only unit tests"""
+    """Run all unit tests including calorie tracking tests"""
     print("\n=== Running Unit Tests ===")
-    unittest.TextTestRunner(verbosity=2).run(unittest.defaultTestLoader.loadTestsFromTestCase(CalTrackUnitTests))
-
-def run_calorie_tracking_tests():
-    """Run only calorie tracking tests"""
-    print("\n=== Running Calorie Tracking Tests ===")
-    unittest.TextTestRunner(verbosity=2).run(unittest.defaultTestLoader.loadTestsFromTestCase(CalorieTrackingTests))
+    # Create a test suite with both unit test classes
+    suite = unittest.TestSuite()
+    suite.addTest(unittest.defaultTestLoader.loadTestsFromTestCase(CalTrackUnitTests))
+    suite.addTest(unittest.defaultTestLoader.loadTestsFromTestCase(CalorieTrackingTests))
+    unittest.TextTestRunner(verbosity=2).run(suite)
 
 def run_selenium_login_page_test():
     """Run only the login page test"""
@@ -84,27 +125,19 @@ def run_all_tests():
     """Run all tests"""
     print("\n=== Running All Tests ===")
     
-    # Run unit tests
-    print("\n--- Running Unit Tests ---")
-    unittest.TextTestRunner(verbosity=2).run(
-        unittest.defaultTestLoader.loadTestsFromTestCase(CalTrackUnitTests)
-    )
-    
-    # Run calorie tracking tests
-    print("\n--- Running Calorie Tracking Tests ---")
-    unittest.TextTestRunner(verbosity=2).run(
-        unittest.defaultTestLoader.loadTestsFromTestCase(CalorieTrackingTests)
-    )
+    # Run unit tests (including calorie tracking tests)
+    run_unit_tests()
     
     # Run selenium tests one by one
     run_all_selenium_tests()
 
 if __name__ == "__main__":
     if len(sys.argv) > 1:
-        if sys.argv[1] == 'unit':
+        if sys.argv[1] in ['-h', '--help']:
+            print(__doc__)
+            sys.exit(0)
+        elif sys.argv[1] == 'unit':
             run_unit_tests()
-        elif sys.argv[1] == 'calorie':
-            run_calorie_tracking_tests()
         elif sys.argv[1] == 'selenium':
             run_all_selenium_tests()
         elif sys.argv[1] == 'login_page':
@@ -120,6 +153,8 @@ if __name__ == "__main__":
         elif sys.argv[1] == 'view_profile':
             run_selenium_view_profile_test()
         else:
-            print("Unknown test suite. Available options: unit, calorie, selenium, login_page, login_success, login_failure, registration, track_exercise, view_profile")
+            print("Unknown test suite.")
+            print("Available options: unit, selenium, login_page, login_success, login_failure, registration, track_exercise, view_profile")
+            print("Use -h or --help for more information.")
     else:
         run_all_tests()
